[Sata] Fix.

+++ source/Cosmos.HAL2/BlockDevice/Ports/Sata.cs
@@ -33,7 +33,7 @@ namespace Cosmos.HAL.BlockDevice.Ports
                 return;
             }
 
-            Mem = new Core.MemoryGroup.AHCI(RegularSectorSize);
+            Mem = new Core.MemoryGroup.AHCI((uint)RegularSectorSize);
 
             mPortReg = aSATAPort;
